Return segmentation mask as uint8.

+++ paulssonlab/src/paulssonlab/shenker/matriarch/trench_segmentation/watershed.py
@@ -66,4 +66,4 @@ def segment_trench(img, threshold=4e-2, diagnostics=None):
     )
     if diagnostics is not None:
         diagnostics["watershed_labels_eroded"] = _trench_img(watershed_labels_eroded)
-    return watershed_labels_eroded  # _rescaled
+    return watershed_labels_eroded.astype(np.uint8)
